Rename all namespaces Microsoft.DotNet.Core => Microsoft.NETCore.
