+++ src/Tasks/Microsoft.NETCore.Build.Tasks/FileGroup.cs
@@ -6,7 +6,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 
-namespace Microsoft.DotNet.Core.Build.Tasks
+namespace Microsoft.NETCore.Build.Tasks
 {
     /// <summary>
     /// Values for File Group Metadata corresponding to the groups in a target library
