+++ src/Tasks/Microsoft.NETCore.Build.Tasks/GenerateDepsFile.cs
@@ -9,7 +9,7 @@ using Newtonsoft.Json;
 using NuGet.Frameworks;
 using NuGet.ProjectModel;
 
-namespace Microsoft.DotNet.Core.Build.Tasks
+namespace Microsoft.NETCore.Build.Tasks
 {
     /// <summary>
     /// Generates the $(project).deps.json file.
