+++ src/Tasks/Microsoft.NETCore.Build.Tasks/LockFileTargetExtensions.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using NuGet.Packaging.Core;
 using NuGet.ProjectModel;
 
-namespace Microsoft.DotNet.Core.Build.Tasks
+namespace Microsoft.NETCore.Build.Tasks
 {
     internal static class LockFileTargetExtensions
     {
