+++ src/Tasks/Microsoft.NETCore.Build.Tasks/ResolvePublishAssemblies.cs
@@ -9,7 +9,7 @@ using NuGet.Configuration;
 using NuGet.Frameworks;
 using NuGet.ProjectModel;
 
-namespace Microsoft.DotNet.Core.Build.Tasks
+namespace Microsoft.NETCore.Build.Tasks
 {
     /// <summary>
     /// Resolves the assemblies to be published for a .NET app.
