+++ src/Tasks/Microsoft.NETCore.Build.Tasks/RuntimeOptions.cs
@@ -5,7 +5,7 @@ using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace Microsoft.DotNet.Core.Build.Tasks
+namespace Microsoft.NETCore.Build.Tasks
 {
     internal class RuntimeOptions
     {
