+++ test/Microsoft.NETCore.TestFramework/Assertions/DirectoryInfoAssertions.cs
@@ -10,7 +10,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Microsoft.DotNet.TestFramework.Assertions
+namespace Microsoft.NETCore.TestFramework.Assertions
 {
     public class DirectoryInfoAssertions
     {
