+++ test/Microsoft.NETCore.TestFramework/Assertions/DirectoryInfoExtensions.cs
@@ -7,7 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Microsoft.DotNet.TestFramework.Assertions
+namespace Microsoft.NETCore.TestFramework.Assertions
 {
     public static class DirectoryInfoExtensions
     {
